requirements.txt
amqp==5.0.2
appdirs==1.4.3
billiard==3.6.3.0
CacheControl==0.12.6
celery==5.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Django==2.2.13
django-filter==2.3.0
django-redis==4.12.1
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
Jinja2==2.11.2
kombu==5.0.2
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
prompt-toolkit==3.0.7
PyJWT==1.7.1
pyparsing==2.4.6
pytoml==0.1.21
pytz==2020.1
redis==3.5.3
requests==2.22.0
retrying==1.3.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
